requirements.txt
beautifulsoup4==4.6.0
bootstrap-admin==0.3.9
certifi==2018.1.18
Django==1.11.8
django-markdown-deux==1.0.5
django-markdownx==2.0.22
django-object-actions==0.10.0
Markdown==2.6.11
markdown2==2.3.5
markdownify==0.4.1
pytz==2018.3
six==1.11.0
